test: nvd datasource
