+++ test/test_nvd_api.py
@@ -48,7 +48,7 @@ class TestNVD_API:
             time_of_last_update=datetime.now() - timedelta(days=4)
         )
         await nvd_api.get()
-        cvedb = CVEDB(cachedir=self.outdir, nvd_type="api")
+        cvedb = CVEDB(cachedir=self.outdir)
         cvedb.all_cve_entries = nvd_api.all_cve_entries
         cvedb.init_database()
         cvedb.populate_db()
