+++ test/test_source_nvd.py
@@ -1,0 +1,56 @@
+import shutil
+import tempfile
+from test.utils import LONG_TESTS
+
+import aiohttp
+import pytest
+
+from cve_bin_tool.sources import nvd
+
+
+class TestSourceNVD:
+    @classmethod
+    def setup_class(cls):
+        cls.nvd = nvd.NVD()
+        cls.nvd.cachedir = tempfile.mkdtemp(prefix="cvedb-")
+
+    @classmethod
+    def teardown_class(cls):
+        shutil.rmtree(cls.nvd.cachedir)
+
+    @pytest.mark.asyncio
+    @pytest.mark.skipif(
+        LONG_TESTS() != 1, reason="Skipping NVD calls due to rate limits"
+    )
+    async def test_00_getmeta(self):
+        connector = aiohttp.TCPConnector(limit_per_host=19)
+        async with aiohttp.ClientSession(
+            connector=connector, trust_env=True
+        ) as session:
+            _jsonurl, meta = await self.nvd.getmeta(
+                session,
+                "https://nvd.nist.gov/feeds/json/cve/1.1/nvdcve-1.1-modified.meta",
+            )
+        assert "sha256" in meta
+
+    @pytest.mark.asyncio
+    @pytest.mark.skipif(
+        LONG_TESTS() != 1, reason="Skipping NVD calls due to rate limits"
+    )
+    async def test_01_cache_update(self):
+        async with aiohttp.ClientSession(trust_env=True) as session:
+            jsonurl, meta = await self.nvd.getmeta(
+                session, "https://nvd.nist.gov/feeds/json/cve/1.1/nvdcve-1.1-2015.meta"
+            )
+            assert "sha256" in meta
+            await self.nvd.cache_update(session, jsonurl, meta["sha256"])
+
+    # @pytest.mark.asyncio
+    # @pytest.mark.skipif(
+    #     LONG_TESTS() != 1, reason="Skipping NVD calls due to rate limits"
+    # )
+    # async def test_03_refresh(self):
+    #     await self.cvedb.refresh()
+    #     years = self.nvd.nvd_years()
+    #     for year in range(2002, datetime.datetime.now().year):
+    #         assert year in years, f"Missing NVD data for {year}"
